+++ CHANGES.md
@@ -70,6 +70,7 @@
 * Support custom coders in Reshuffle ([#29908](https://github.com/apache/beam/issues/29908), [#33356](https://github.com/apache/beam/issues/33356)).
 * [Java] Upgrade SLF4J to 2.0.16. Update default Spark version to 3.5.0. ([#33574](https://github.com/apache/beam/pull/33574))
 * X feature added (Java/Python) ([#X](https://github.com/apache/beam/issues/X)).
+* Added caching for BigQuery table definitions to reduce API calls from workers (Python) ([#34076](https://github.com/apache/beam/issues/34076)).
 
 ## Breaking Changes